esp32c3: Move SWD functions into RTC_CNTL common implementation
