+++ esp32c3-hal/examples/hello_rgb.rs
@@ -46,7 +46,7 @@ fn main() -> ! {
 
     // Disable watchdogs
     rtc_cntl.set_super_wdt_enable(false);
-    rtc_cntl.set_wdt_enable(false);
+    rtc_cntl.set_wdt_global_enable(false);
     wdt0.disable();
 
     // Configure RMT peripheral globally
