+++ esp32c3-hal/examples/read_efuse.rs
@@ -33,7 +33,7 @@ fn main() -> ! {
 
     // Disable watchdog timers
     rtc_cntl.set_super_wdt_enable(false);
-    rtc_cntl.set_wdt_enable(false);
+    rtc_cntl.set_wdt_global_enable(false);
     wdt0.disable();
     wdt1.disable();
