+++ esp32c3-hal/examples/usb_serial_jtag.rs
@@ -35,7 +35,7 @@ fn main() -> ! {
 
     // Disable watchdog timers
     rtc_cntl.set_super_wdt_enable(false);
-    rtc_cntl.set_wdt_enable(false);
+    rtc_cntl.set_wdt_global_enable(false);
     wdt0.disable();
     wdt1.disable();
